requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.5.1
firebase-admin==6.5.0
boto3==1.34.44
python-dotenv==1.0.0
